refactor(design-system): promote SplitButton to its own component directory
