+++ .rhythmguard-baseline.json
@@ -1,5 +1,5 @@
 {
-  "createdAt": "2026-07-03T06:11:52.073Z",
+  "createdAt": "2026-07-03T06:31:34.357Z",
   "directory": "nextjs-app/shared",
   "findings": [
     {
@@ -361,56 +361,6 @@
       "text": "Unexpected raw scale value \"0.5rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
       "type": "token-opportunity",
       "value": "0.5rem"
-    },
-    {
-      "column": 16,
-      "file": "nextjs-app/shared/components/Button/SplitButton.module.css",
-      "key": "rhythmguard/use-scale\u001foff-scale\u001fnextjs-app/shared/components/Button/SplitButton.module.css\u001f112\u001f16\u001f0.125rem\u001fUnexpected off-scale value \"0.125rem\". Use scale values (nearest: 0px or 4px). (rhythmguard/use-scale)",
-      "line": 112,
-      "rule": "rhythmguard/use-scale",
-      "text": "Unexpected off-scale value \"0.125rem\". Use scale values (nearest: 0px or 4px). (rhythmguard/use-scale)",
-      "type": "off-scale",
-      "value": "0.125rem"
-    },
-    {
-      "column": 23,
-      "file": "nextjs-app/shared/components/Button/SplitButton.module.css",
-      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/Button/SplitButton.module.css\u001f61\u001f23\u001f100%\u001fUnexpected raw scale value \"100%\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "line": 61,
-      "rule": "rhythmguard/prefer-token",
-      "text": "Unexpected raw scale value \"100%\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "type": "token-opportunity",
-      "value": "100%"
-    },
-    {
-      "column": 18,
-      "file": "nextjs-app/shared/components/Button/SplitButton.module.css",
-      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/Button/SplitButton.module.css\u001f68\u001f18\u001f0.5rem\u001fUnexpected raw scale value \"0.5rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "line": 68,
-      "rule": "rhythmguard/prefer-token",
-      "text": "Unexpected raw scale value \"0.5rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "type": "token-opportunity",
-      "value": "0.5rem"
-    },
-    {
-      "column": 19,
-      "file": "nextjs-app/shared/components/Button/SplitButton.module.css",
-      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/Button/SplitButton.module.css\u001f69\u001f19\u001f0.75rem\u001fUnexpected raw scale value \"0.75rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "line": 69,
-      "rule": "rhythmguard/prefer-token",
-      "text": "Unexpected raw scale value \"0.75rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "type": "token-opportunity",
-      "value": "0.75rem"
-    },
-    {
-      "column": 16,
-      "file": "nextjs-app/shared/components/Button/SplitButton.module.css",
-      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/Button/SplitButton.module.css\u001f112\u001f16\u001f0.125rem\u001fUnexpected raw scale value \"0.125rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "line": 112,
-      "rule": "rhythmguard/prefer-token",
-      "text": "Unexpected raw scale value \"0.125rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
-      "type": "token-opportunity",
-      "value": "0.125rem"
     },
     {
       "column": 27,
@@ -4241,6 +4191,56 @@
       "text": "Unexpected raw scale value \"-4rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
       "type": "token-opportunity",
       "value": "-4rem"
+    },
+    {
+      "column": 16,
+      "file": "nextjs-app/shared/components/SplitButton/SplitButton.module.css",
+      "key": "rhythmguard/use-scale\u001foff-scale\u001fnextjs-app/shared/components/SplitButton/SplitButton.module.css\u001f112\u001f16\u001f0.125rem\u001fUnexpected off-scale value \"0.125rem\". Use scale values (nearest: 0px or 4px). (rhythmguard/use-scale)",
+      "line": 112,
+      "rule": "rhythmguard/use-scale",
+      "text": "Unexpected off-scale value \"0.125rem\". Use scale values (nearest: 0px or 4px). (rhythmguard/use-scale)",
+      "type": "off-scale",
+      "value": "0.125rem"
+    },
+    {
+      "column": 23,
+      "file": "nextjs-app/shared/components/SplitButton/SplitButton.module.css",
+      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/SplitButton/SplitButton.module.css\u001f61\u001f23\u001f100%\u001fUnexpected raw scale value \"100%\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "line": 61,
+      "rule": "rhythmguard/prefer-token",
+      "text": "Unexpected raw scale value \"100%\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "type": "token-opportunity",
+      "value": "100%"
+    },
+    {
+      "column": 18,
+      "file": "nextjs-app/shared/components/SplitButton/SplitButton.module.css",
+      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/SplitButton/SplitButton.module.css\u001f68\u001f18\u001f0.5rem\u001fUnexpected raw scale value \"0.5rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "line": 68,
+      "rule": "rhythmguard/prefer-token",
+      "text": "Unexpected raw scale value \"0.5rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "type": "token-opportunity",
+      "value": "0.5rem"
+    },
+    {
+      "column": 19,
+      "file": "nextjs-app/shared/components/SplitButton/SplitButton.module.css",
+      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/SplitButton/SplitButton.module.css\u001f69\u001f19\u001f0.75rem\u001fUnexpected raw scale value \"0.75rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "line": 69,
+      "rule": "rhythmguard/prefer-token",
+      "text": "Unexpected raw scale value \"0.75rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "type": "token-opportunity",
+      "value": "0.75rem"
+    },
+    {
+      "column": 16,
+      "file": "nextjs-app/shared/components/SplitButton/SplitButton.module.css",
+      "key": "rhythmguard/prefer-token\u001ftoken-opportunity\u001fnextjs-app/shared/components/SplitButton/SplitButton.module.css\u001f112\u001f16\u001f0.125rem\u001fUnexpected raw scale value \"0.125rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "line": 112,
+      "rule": "rhythmguard/prefer-token",
+      "text": "Unexpected raw scale value \"0.125rem\". Use design tokens for scale decisions. (rhythmguard/prefer-token)",
+      "type": "token-opportunity",
+      "value": "0.125rem"
     },
     {
       "column": 8,
