+++ nextjs-app/shared/components/Button/index.ts
@@ -5,5 +5,3 @@ export type {
   ButtonProps,
   ButtonSurface,
 } from "./Button";
-export { default as SplitButton } from "./SplitButton";
-export type { SplitButtonProps } from "./SplitButton";
