+++ nextjs-app/shared/components/SplitButton/SplitButton.mdx
@@ -1,0 +1,17 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+import * as Stories from './SplitButton.stories.tsx';
+
+<Meta of={Stories} />
+
+# SplitButton
+
+**Status:** alpha · **Tier:** molecule · **Group:** interaction
+
+A primary action button paired with a dropdown toggle for related secondary
+actions (e.g. Save / Save as, Export PDF / Export CSV), with collision-aware
+menu placement and optional nested submenus. This page is an honest alpha
+stub: SplitButton was just promoted out of `Button/` into its own directory
+as a structural move, and has not yet been through the a11y review, token
+audit, or forced-colors/light-dark verification that beta promotion requires.
+See `docs/superpowers/specs/2026-07-03-astryx-level-design-system-roadmap.md`
+for the tracked follow-up work.
